buttontext fixed, max birth date, refactor css
