+++ src/Components/Button.js
@@ -2,8 +2,7 @@ import React, { useContext } from "react";
 import { AppContext } from "./AppProvider";
 import './Styling/Button.css';
 
-/* A general Button for consistency throughout layout. Takes an onClick & buttonText props. Right now the buttonTexxt is not 
-implemented fully am working on this.*/ 
+/* A general Button for consistency throughout layout. Takes an onClick & buttonText props.*/ 
 
 const DefaultButton = (props) => {
   const { onClick, buttonText } = props;
